DocumentLite rename, Use real global
