+++ src/worker-thread/index.nodom.amp.ts
@@ -18,7 +18,7 @@ import { HydrateFunction } from './hydrate';
 import { AMP } from './amp/amp';
 import { callFunctionMessageHandler, exportFunction } from './function';
 import { WorkerNoDOMGlobalScope } from './WorkerDOMGlobalScope';
-import { DocumentStub } from './dom/DocumentLite';
+import { DocumentStub } from './dom/DocumentStub';
 import { deleteGlobals } from './amp/delete-globals';
 import { initializeStorage } from './initialize-storage';
 import { WorkerStorageInit } from './initialize-storage';
